+++ test/unit-tests/function/statistics/sum.test.js
@@ -27,17 +27,13 @@ describe('sum', function () {
     assert.strictEqual(sum([['2', '3'], ['4', '5']]), 14)
   })
 
-  // eslint-disable-next-line mocha/no-skipped-tests
-  it.skip('should return the max of strings by their numerical value (with BigNumber config)', function () {
-    // TODO: requires math.add to recon with config.number when parsing strings
+  it('should return the sum of strings (with BigNumber config)', function () {
     const bigmath = math.create({ number: 'BigNumber' })
     assert.deepStrictEqual(bigmath.sum('10', '3', '4', '2'), bigmath.bignumber('19'))
     assert.deepStrictEqual(bigmath.sum('10'), bigmath.bignumber(10))
   })
 
-  // eslint-disable-next-line mocha/no-skipped-tests
-  it.skip('should return the max of strings by their numerical value (with bigint config)', function () {
-    // TODO: requires math.add to recon with config.number when parsing strings
+  it('should return the sum of strings (with bigint config)', function () {
     const bigmath = math.create({ number: 'bigint' })
     assert.strictEqual(bigmath.sum('10', '3', '4', '2'), 19n)
     assert.strictEqual(bigmath.sum('10'), 10n)
@@ -134,7 +130,7 @@ describe('sum', function () {
     assert.throws(function () { sum(new Date(), 2) }, /Cannot calculate sum, unexpected type of argument/)
     assert.throws(function () { sum(2, 3, null) }, /Cannot calculate sum, unexpected type of argument/)
     assert.throws(function () { sum([2, 3, null]) }, /Cannot calculate sum, unexpected type of argument/)
-    assert.throws(function () { sum('a', 'b') }, /Error: Cannot convert "a" to a number/)
+    assert.throws(function () { sum('a', 'b') }, /SyntaxError: String "a" is not a valid number/)
     assert.throws(function () { sum('a') }, /SyntaxError: String "a" is not a valid number/)
   })