+++ test/unit-tests/function/statistics/sum.test.ts
@@ -41,9 +41,7 @@ describe('sum', function () {
     )
   })
 
-  // eslint-disable-next-line mocha/no-skipped-tests
-  it.skip('should return the max of strings by their numerical value (with BigNumber config)', function () {
-    // TODO: requires math.add to recon with config.number when parsing strings
+  it('should return the sum of strings (with BigNumber config)', function () {
     const bigmath = math.create({ number: 'BigNumber' })
     assert.deepStrictEqual(
       bigmath.sum('10', '3', '4', '2'),
@@ -52,9 +50,7 @@ describe('sum', function () {
     assert.deepStrictEqual(bigmath.sum('10'), bigmath.bignumber(10))
   })
 
-  // eslint-disable-next-line mocha/no-skipped-tests
-  it.skip('should return the max of strings by their numerical value (with bigint config)', function () {
-    // TODO: requires math.add to recon with config.number when parsing strings
+  it('should return the sum of strings (with bigint config)', function () {
     const bigmath = math.create({ number: 'bigint' })
     assert.strictEqual(bigmath.sum('10', '3', '4', '2'), 19n)
     assert.strictEqual(bigmath.sum('10'), 10n)
@@ -223,7 +219,7 @@ describe('sum', function () {
     }, /Cannot calculate sum, unexpected type of argument/)
     assert.throws(function () {
       sum('a', 'b')
-    }, /Error: Cannot convert "a" to a number/)
+    }, /SyntaxError: String "a" is not a valid number/)
     assert.throws(function () {
       sum('a')
     }, /SyntaxError: String "a" is not a valid number/)